requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-ollama>=0.0.1

# 向量数据库
chromadb>=0.4.22
faiss-cpu>=1.7.4

# 文档处理
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.4
markdown>=3.5.1

# 数据可视化
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# 邮件处理
imaplib2>=3.6
python-email>=0.2.0
smtplib>=0.0.1

# 日历处理
icalendar>=5.0.11

# Web框架
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6

# 任务调度
apscheduler>=3.10.4

# 工具库
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.1

# 日志
loguru>=0.7.2

# 安全
cryptography>=41.0.7

# 测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.26.0

# 代码质量
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
